python-prompt-generator/requirements.txt
anthropic==0.39.0
python-dotenv==1.0.0
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
pydantic==2.5.0
Pillow==10.2.0
typing-extensions==4.9.0
openai==1.3.0
numpy==1.24.0
matplotlib==3.7.0